Idea for reorganization
An idea for organizing harald to conform more to how the spec is
organized. The hope is to also make adding support for a command/event
as easy as just adding a new module that matches some behaviour
expectations.

This only applies this idea to `Read_Local_Name` command from
`ControllerAndBaseband` section. If we like this idea, we can move
forward with the organization